chore: upgrade kafkaflow.sample to modern c# practices
