+++ samples/KafkaFlow.Sample/TestMessage.cs
@@ -1,0 +1,10 @@
+using System.Runtime.Serialization;
+
+namespace KafkaFlow.Sample;
+
+[DataContract]
+public class TestMessage
+{
+    [DataMember(Order = 1)] 
+    public string Text { get; set; }
+}
