+++ website/docs/getting-started/samples.md
@@ -1,0 +1,39 @@
+---
+sidebar_position: 5
+---
+
+# Samples
+
+We know that working code is a valuable learning tool for many, so here you can find a list of samples built to demonstrate KafkaFlow capabilities.
+
+
+## Basic
+This is a simple sample that shows how to produce and consume messages.
+
+You can find the code here: [/samples/KafkaFlow.Sample](https://github.com/Farfetch/kafkaflow/tree/master/samples/KafkaFlow.Sample)
+
+## Batching
+This is a sample that shows batch processing using KafkaFlow.
+
+You can find the code here: [/samples/KafkaFlow.Sample.BatchOperations](https://github.com/Farfetch/kafkaflow/tree/master/samples/KafkaFlow.Sample.BatchOperations)
+
+## Schema Registry
+This is a sample that shows how to use Schema Registry with KafkaFlow.
+
+You can find the code here: [/samples/KafkaFlow.Sample](https://github.com/Farfetch/kafkaflow/tree/master/samples/KafkaFlow.Sample.SchemaRegistry)
+
+
+## Web API
+This sample shows how to host an administration Web API for administrative operations.
+
+You can find the code here: [/samples/KafkaFlow.Sample](https://github.com/Farfetch/kafkaflow/tree/master/samples/KafkaFlow.Sample.WebApi)
+
+## Dashboard
+This sample shows how to use KafkaFlow to expose an administration Dashboard.
+
+You can find the code here: [/samples/KafkaFlow.Sample](https://github.com/Farfetch/kafkaflow/tree/master/samples/KafkaFlow.Sample.Dashboard)
+
+## Flow of Control
+This is a sample that shows how to control the state of a consumer (Starting, Pausing, Stopping, etc.) programmatically.
+
+You can find the code here: [/samples/KafkaFlow.Sample.FlowControl](https://github.com/Farfetch/kafkaflow/tree/master/samples/KafkaFlow.Sample.FlowControl)
